feat(shopinbit): create DesktopServicesView container
